feat(app): add ProjectsController and ProjectsService for project management functionality fix(app.module.ts): update module imports and providers to include new ProjectsController and ProjectsService fix(settings.controller.ts): add update and create methods for settings management fix(settings.service.ts): implement update method for settings service refactor(project.guard.ts): update import path for ProjectsService to reflect new structure chore(remove): delete unused dtos directory and clean up project structure
