+++ server/src/app/controllers/settings.controller.ts
@@ -1,7 +1,8 @@
 import {SettingsService} from '@/app/services/settings.service';
 import {SettingDto} from '@/entities/setting.entity';
 import {Response} from '@/scaffold/decorators/response';
-import {Controller, Get} from '@nestjs/common';
+import {Body, Controller, Get, Put} from '@nestjs/common';
+import {Patch} from '@nestjs/common/decorators/http/request-mapping.decorator';
 import {ApiOperation, ApiTags} from '@nestjs/swagger';
 
 @ApiTags('App')
@@ -16,5 +17,19 @@ export class SettingsController {
     @Response(SettingDto)
     public async get(): Promise<SettingDto> {
         return await this.settings.get(1);
+    }
+
+    @Put()
+    @ApiOperation({summary: `Replace app settings`})
+    @Response(SettingDto)
+    public async create(@Body() data: SettingDto): Promise<SettingDto> {
+        return await this.settings.update(1, data);
+    }
+
+    @Patch()
+    @ApiOperation({summary: `Update app settings`})
+    @Response(SettingDto)
+    public async update(@Body() data: SettingDto): Promise<SettingDto> {
+        return await this.settings.update(1, data);
     }
 }
