+++ server/src/projects/projects.module.ts
@@ -1,5 +1,5 @@
+import {AppModule} from '@/app/app.module';
 import {ProjectEntity} from '@/entities/project.entity';
-import {ProjectsController} from '@/projects/controllers/projects.controller';
 import {WORKSPACES_REPOSITORY} from '@/projects/project-symbols';
 import {ProjectDatabasesService} from '@/projects/services/project-databases.service';
 import {ProjectEntityService} from '@/projects/services/project-entity.service';
@@ -7,22 +7,24 @@ import {
     PROJECT_EXTENSION,
     ProjectsStorageService
 } from '@/projects/services/projects-storage.service';
-import {ProjectsService} from '@/projects/services/projects.service';
 import {UtilsModule} from '@/utils/utils.module';
 import {Module, Scope} from '@nestjs/common';
 import {TypeOrmModule} from '@nestjs/typeorm';
 import {ProjectGuard} from './gaurds/project.guard';
 
 @Module({
-    imports: [TypeOrmModule.forFeature([ProjectEntity]), UtilsModule],
-    controllers: [ProjectsController],
+    imports: [
+        AppModule,
+        TypeOrmModule.forFeature([ProjectEntity]),
+        UtilsModule
+    ],
+    controllers: [],
     providers: [
         ProjectGuard,
         {
             provide: PROJECT_EXTENSION,
             useValue: '.aigraphr'
         },
-        ProjectsService,
         ProjectsStorageService,
         ProjectDatabasesService,
         {
@@ -36,7 +38,6 @@ import {ProjectGuard} from './gaurds/project.guard';
     ],
     exports: [
         ProjectGuard,
-        ProjectsService,
         ProjectsStorageService,
         ProjectDatabasesService,
         ProjectEntityService,
